// client-app/src/app/models/activity.ts
export interface IActivity {
  id: string;
  title: string;
  description: string;
  date: Date;
  category: string;
  city: string;
  venue: string;
  isHost:boolean;
  isGoing:boolean;
  attendees: IAttendee[];
  comments: IComment[];
}

export interface IComment {
    id: string;
    createdAt: Date;
    body: string;
    username: string;
    displayName: string;
    image: string;
}

export interface IActivityFormValues extends Partial<IActivity> {
  time?: Date;
}
export class ActivityFormValues implements IActivityFormValues {
  id?: string = undefined;
  title: string = "";
  description: string = "";
  date?: Date = undefined;
  time?: Date = undefined;
  category: string = "";
  city: string = "";
  venue: string = "";
  constructor(init?: IActivityFormValues) {
    if(init && init.date){
      init.time = init.date;
    }
    Object.assign(this,init);
  }
  
}
export interface IAttendee{
  username:string;
  displayName:string;
  image:string;
  isHost:boolean;
}